Refactor imports for relative paths and module exposure
Updated various modules to use relative imports for better package structure and maintainability. Exposed key functions and classes in new_updater's __init__.py for easier access. Adjusted test scripts and main scripts to reflect new import paths and usage patterns. Minor code cleanups and improved argument parsing in med.py.
