+++ newupdater/medask.py
@@ -9,26 +9,19 @@ python3 core8/pwb.py newupdater/medask -start:!
 python3 core8/pwb.py newupdater/medask -ns:0 -usercontribs:Edoderoobot
 python3 core8/pwb.py newupdater/medask -ns:0 -usercontribs:Ghuron
 """
-#
-# (C) Ibrahem Qasim, 2023
-#
-#
+
 import sys
 import urllib
 import urllib.parse
-
-# ---
 from pathlib import Path
 
-Dir = Path(__file__).parent
-# ---
-sys.path.append(str(Dir))
+sys.path.append(str(Path(__file__).parent))
 # ---
 from newapi import printe
 from newapi.mdwiki_page import NEW_API
 
 from apis import mdwiki_api
-from new_updater.MedWorkNew import work_on_text
+from new_updater import work_on_text
 
 import mdapi
 # ---
