+++ newupdater/new_updater/MedWorkNew.py
@@ -6,15 +6,15 @@ python3 pwb.py newupdater/med Retinol from_toolforge
 """
 import re
 # ---
-from new_updater.bots import expend  # expend_infoboxs_and_fix(text)
-from new_updater.bots import expend_new  # expend_infoboxs(text)
-from new_updater.bots import old_params
+from .bots import expend  # expend_infoboxs_and_fix(text)
+from .bots import expend_new  # expend_infoboxs(text)
+from .bots import old_params
 
-from new_updater.helps import echo_debug
-from new_updater import mv_section  # mv_section.move_External_links_section
-from new_updater import drugbox  # drugbox.TextProcessor
-from new_updater import resources_new
-from new_updater import chembox  # fix_Chembox
+from .helps import echo_debug
+from .mv_section import move_External_links_section
+from .drugbox import TextProcessor
+from .resources_new import move_resources
+from .chembox import fix_Chembox
 # ---
 lkj = r"<!--\s*(Monoclonal antibody data|External links|Names*|Clinical data|Legal data|Legal status|Pharmacokinetic data|Chemical and physical data|Definition and medical uses|Chemical data|Chemical and physical data|index_label\s*=\s*Free Base|\w+ \w+ data|\w+ \w+ \w+ data|\w+ data|\w+ status|Identifiers)\s*-->"
 # ---
@@ -27,7 +27,7 @@ def drugbox_work(new_text, text):
     # ---
     # new_text = re.sub(r'<!--\s*\|\s*type\s*=\s*mab\s*\/\s*vaccine\s*\/\s*combo\s*-->', '<!-- type = mab / vaccine / combo -->', new_text, flags=re.IGNORECASE)
     # ---
-    bot = drugbox.TextProcessor(new_text)
+    bot = TextProcessor(new_text)
     # ---
     drugbox_text = bot.get_old_temp()
     drug_box_new = bot.get_new_temp()
@@ -66,11 +66,11 @@ def work_on_text_md(title, text):
     # ---
     new_text = old_params.rename_params(new_text)
     # ---
-    new_text = resources_new.move_resources(new_text, title, lkj=lkj, lkj2=lkj2)
+    new_text = move_resources(new_text, title, lkj=lkj, lkj2=lkj2)
     # ---
     new_text = drugbox_work(new_text, text)
     # ---
-    bot2 = mv_section.move_External_links_section(new_text)
+    bot2 = move_External_links_section(new_text)
     # ---
     new_text = bot2.make_new_txt()
     # ---
@@ -88,7 +88,7 @@ def work_on_text(title, text):
     Chem = re.search(r"{{(Chembox)", newtext, flags=re.IGNORECASE)
     # ---
     if Chem:
-        bot = chembox.fix_Chembox(newtext)
+        bot = fix_Chembox(newtext)
         newtext = bot.run()
     # ---
     rea = re.search(r"{{(Infobox drug|Drugbox)", newtext, flags=re.IGNORECASE)
