+++ newupdater/new_updater/bots/expend.py
@@ -2,8 +2,8 @@
 from .bots.expend import expend_infoboxs_and_fix
 """
 import wikitextparser as wtp
-from new_updater.lists.expend_lists import IMC_params, dup_params
-from new_updater.helps import echo_debug
+from ..lists.expend_lists import IMC_params, dup_params
+from ..helps import echo_debug
 
 
 def expend_infoboxs_and_fix(new_text):
