+++ newupdater/new_updater/bots/expend_new.py
@@ -2,9 +2,8 @@
 from .bots.expend_new import expend_infoboxs
 """
 import wikitextparser as wtp
-from new_updater.helps import echo_debug
-
-from new_updater.lists.expend_lists import main_temps_list
+from ..helps import echo_debug
+from ..lists.expend_lists import main_temps_list
 
 
 def expend_new(template, min_len=1):
