+++ newupdater/new_updater/bots/old_params.py
@@ -3,7 +3,7 @@ from .bots.old_params import rename_params
 """
 # ---
 import wikitextparser as wtp
-from new_updater.helps import echo_debug
+from ..helps import echo_debug
 
 
 def rename_params(temptext):
@@ -61,4 +61,3 @@ def rename_params(temptext):
         new_temptext = new_temptext.replace(old_temp, temp.string)
     # ---
     return new_temptext
-
